frontend/requirements.txt
antlr4-python3-runtime
python-dateutil
pandas
pandas-highcharts
flask
sqlalchemy
python3-memcached
PyYAML
elasticsearch-dsl>=5.0.0,<6.0.0
hawkular-client
redis
cassandra-driver
